agent/requirements.txt
python-dotenv==1.0.0
uvloop==0.19.0
openai==1.76.2
anthropic==0.40.0
google-generativeai==0.8.3
mcp==1.0.0
httpx==0.27.0
aiohttp==3.9.5
redis==5.0.1
scikit-learn==1.3.0
watchdog==3.0.0
asyncpg==0.29.0
PyYAML==6.0.1
tiktoken==0.5.2